MG232 (Update FDO connection cache)
Fix compiling errors in Linux.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@544 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Services/Feature/ServerFeatureUtil.cpp
@@ -541,7 +541,7 @@ bool MgServerFeatureUtil::ValidatePoolEntry(INT32 readerId)
 {
     bool isValid = false;
 
-    MgServerFeatureProcessor* featureProcessor = ((MgServerFeatureProcessor*)(INT64)readerId);
+    MgServerFeatureProcessor* featureProcessor = (MgServerFeatureProcessor*)readerId;
     if (featureProcessor != NULL)
     {
         MgServerFeatureProcessorType fpType = featureProcessor->GetProcessorType();
